Optimización de archivos
Se ha simplificado la estructura de archivos y se han eliminado los archivos demo desarrollados hasta la fecha, con el objetivo de enfocar el proyecto a una aplicación real.

Ahora el archivo Crew unifica el código de las tres modalidades y el archivo __main__.py comprende los antiguos archivos main, main_demo, main_demo_groq y main_demo_cerebras
